dpif-netdev: Use lockless queue to manage offloads
The dataplane threads (PMDs) send offloading commands to a dedicated
offload management thread.  The current implementation uses a lock
and benchmarks shows a high contention on the queue on high load.

With high-contention, the mutex will more often lead to the locking
thread yield in wait, using a syscall.  This is an issue in the
dataplane.

The mpsc-queue can be used instead.  It uses less cycles and has
lower latency.  Benchmarks shows better behavior both in the case
of a single PMD writing and a single offload thread reading, or
multiple PMDs writing to the queue after polling multiple RXQs.

One trade-off with the new scheme however is to be forced to poll
the queue from the offload thread.  Without mutex, a cond_wait
cannot be used to signal work is available.  The offload thread is
implementing an exponential backoff and will sleep in short increments
when no data is available.  This makes the thread yield, at the price
of higher latency to manage offloads right after an inactivity period.

Change-Id: I1b693c0c2efa669a36462c94c09929d50582fe5e
Signed-off-by: Gaetan Rivet <grive@u256.net>