dpif-netdev: Make megaflow and mark mappings thread objects
In the next commit hardware offloads are managed in several threads.
Each offload is managed by a thread determined by its flow's 'mega_ufid'.

As megaflow to mark and mark to flow mappings are 1:1 and 1:N
respectively, then a single mark exists for a single 'mega_ufid', and
multiple flows uses the same 'mega_ufid'. Because the managing thread will
be choosen using the 'mega_ufid', then each mapping does not need to be
shared with other offload threads.

The mappings are kept as cmap as upcalls can be made that will query them before
enqueuing orders to the offload threads.

To prepare this change, move the mappings within the offload thread
structure.

Change-Id: Id679574f8d51b1ab5371dcc4d617c353958508aa
Signed-off-by: Gaetan Rivet <grive@u256.net>