netdev-offload: Add multi-thread API
Expose functions reporting user configuration of offloading threads, as
well as utility functions for multithreading.

This will only expose the configuration knob to the user, while no
datapath will implement the multiple thread request.

This will allow implementations to use this API for offload thread
management in relevant layers before enabling the actual dataplane
implementation.

The offload thread ID is lazily allocated and can as such be in a
different order than the offload thread start sequence.

The RCU thread will sometime access hardware-offload objects from
a provider for reclamation purposes.  In such case, it will get
a default offload thread ID of 0, that is safe to use as all offload
threads will be quiescent.

Change-Id: I93d6bc63214899323e55dbc91be831858cf21296
Signed-off-by: Gaetan Rivet <grive@u256.net>