netdev-offload: Add function to read hardware offload stats
Add a function for an offload provider to report a device
hardware offload count.  It is not for reporting offload usage
statistics, i.e. in terms of packets or bytes matched by offloads,
but only to know how many offloads are currently programmed into the
device.

Because it is not related to any specific flow, this count is not
integrated with the more general flow statistics.

Change-Id: I59564142d2248041e31d4b7832fd5dd113994cc0
Signed-off-by: Gaetan Rivet <grive@u256.net>