netdev-offload-dpdk: Implement hw-offload statistics read
In the DPDK offload provider, keep track of inserted rte_flow and report
it when asked.  Only one thread updates the counter so consistency is
achieved.  A different thread will query the counter, meaning that the
result is only a moment snapshot that can be incorrect even as it is
read.  On CPU offering cache coherence the counter will eventually
become consistent with the reality.

Change-Id: I995e288c1ad55d8e2f98753059d8e6926a598be8
Signed-off-by: Gaetan Rivet <grive@u256.net>