seq-pool: Use lockless ring as user cache
Remove the per-cache mutex and use the lockless ring as the
cache queue instead.  As such, ID allocation and free are still
thread-safe when sharing user IDs between threads, as well as
when a user steals IDs from another user cache when the pool is
empty.

The deadlock mitigations can be avoided with the lockless structure.
Allocation is more robust as the pool won't fail sometimes on ID
allocation when there are still a few IDs left in another user cache,
due to interlocking.

Performance is roughly the same, but slightly faster with the lockless
ring.

Change-Id: If484fde19672142ab61981950c05e23e29d97c31
Signed-off-by: Gaetan Rivet <grive@u256.net>