netdev-offload-dpdk: Lock rte_flow map access
Add a lock to access the ufid to rte_flow map.  This will protect it
from concurrent write accesses when multiple threads attempt it.

At this point, the reason for taking the lock is not to fullfill the
needs of the DPDK offload implementation anymore. Rewrite the comments
to reflect this change. The lock is still needed to protect against
changes to netdev port mapping.

Change-Id: I2db633570f9e89d060b9f6f3276ce63a3989dfad
Signed-off-by: Gaetan Rivet <grive@u256.net>