conntrack: Use mpsc-queue to store conn expirations
Change the connection expiration lists from ovs_list to mpsc-queue.
This is a pre-step towards reducing the granularity of 'ct_lock'.

It simplifies the responsibilities toward updating the expiration queue.
The dataplane now appends the new conn for expiration once during
creation.  Any further update will only consist in writing the conn
expiration limit and marking the conn for expiration rescheduling.

The ageing thread 'ct_clean' is the only one consuming the expiration
lists.  If a conn was marked for rescheduling by a dataplane, it will
move the conn to the end of the queue.

Once the locks have been reworked, it means neither the dataplane
threads nor 'ct_clean' have to take a lock to update the expiration
lists (assuming the consumer lock is perpetually held by 'ct_clean');

Signed-off-by: Gaetan Rivet <grive@u256.net>
Reviewed-by: Eli Britstein <elibr@nvidia.com>