dpif-netdev: Execute flush from offload thread
When a port is deleted, its offloads must be flushed.  The operation
runs in the thread that initiated it.  Offload data is thus accessed
jointly by the port deletion thread(s) and the offload thread, which
complicates the data access model.

To simplify this model, as a pre-step toward introducing parallel
offloads, execute the flush operation in the offload thread.

Signed-off-by: Gaetan Rivet <grive@u256.net>
Reviewed-by: Maxime Coquelin <maxime.coquelin@redhat.com>